+++ src/core/ReflMngr.cpp
@@ -344,7 +344,7 @@ namespace Ubpa::UDRefl::details {
 }
 
 ReflMngr::ReflMngr() {
-	RegisterType(Global, 0, 0);
+	RegisterType(GlobalType, 0, 0);
 
 	RegisterType<ContainerType>();
 	AddField<ContainerType::Array>("Array");