+++ src/test/10_dynamic/main.cpp
@@ -12,13 +12,14 @@ enum class Color {
 };
 
 int main() {
-	Mngr.AddDynamicField<const std::string>(Global, "author", "Ubpa");
-	Mngr.AddDynamicField<const Color>(Global, "theme", Color::Red);
-	Mngr.RegisterType<int>();
+	Mngr.AddDynamicField<const std::string>(GlobalType, "author", "Ubpa");
+	Mngr.AddDynamicField<const Color>(GlobalType, "theme", Color::Red);
 	Mngr.AddDynamicField<const size_t>(Type_of<int>, "bits", sizeof(int) * 8);
 
-	std::cout << Mngr.Var(Global, "author") << std::endl;
-	std::cout << (Mngr.Var(Global, "theme") == Color::Red) << std::endl;
+	std::cout << Global.Var("author") << std::endl;
+	std::cout << (Global.Var("theme") == Color::Red) << std::endl;
 
-	std::cout << Mngr.Var(Type_of<int>, "bits") << std::endl;
+	std::cout << ObjectView_of<int>.Var("bits") << std::endl;
+
+	return 0;
 }