+++ doc/change_log.md
@@ -7,6 +7,7 @@
 - 0.8.2
   - use concept for traits (operations, container API and flag)
   - let some APIs return fundamental type (`void`, `std::size_t`, `bool`)
+  - original `constexpr Type Global` rename to `GlobalType`, add `constexpr ObjectView Global` for convenience
 - 0.8.1
   - support
     - pointer, array